Partial fix for Issue #30 on scoping-dev branch.
The trouble is that some special forms such as 'if' are baked into the core evaluator. So they circumvent the lexical scoping rules because their pattern is matched prior to reaching the pattern that for general function application. This fix extracts that pattern into a separate function, and adds logic to detect if it needs to be called into from the 'if' special forms.

The fix needs to be generalized to take care of all the special forms, but for now it is good enough to fix two of the issues from this ticket (2 instead of 6) and (10 instead of 111).