Added logic for (set!)
When a set! symbol is found during expansion, the code checks to see if the variable exists in the lexical environment but not within the macro. In this case only, a rename var will be inserted into the rename env so that the macro knows that the symbol is actually a bound variable. Otherwise the rest of the macro logic will pick up the var, so nothing needs to be done for set!