Work around for issue #95
Another issue with load is that it does not handle the case when code is passed as the filename. This is exactly the scenario for require-extension. For now we can work around it by calling `evalLisp` to evaluate the code at compile time. For basic use cases this works well enough. The changes need to be generalized and extended to the other versions of `load`.