Added back stdlib test suite
The issue here is that the compiler does not allow a reference to a variable that has not been defined yet, whereas the interpreter does not care. For now an extra `define` statement will just be added to the test suite. Allowing this behavior in the compiler might be considered at a future time.