Issue #111 - Optimize calls containing vars / lits
Allow optimizing a function call that does not contain any function calls as an argument. Variables are handled by collecting a list of them along with the correponding symbol to use in the Haskell code for a temporary variable. A call to getRtVar is then used to extract each one from the env prior to calling apply.