Issue #113 - Cache macro expansions in funcs
Cache macro expansions when defining function bodies, as per the corresponding code in Core. This prevents issues when evaluating said bodies during an ER macro transformation, where a special form could otherwise accidentally end up as the function that is being evaluated. This would result in an "unbound variable" error since the special form is not defined - IE, "if is not bound".