Fix for issue #121
The root problem here was that the module environment was not being loaded for a module that had already been encountered by the compiler and loaded into memory. This is why a bogus 'value' parameter was encountered, since the compiler never bothered to update it to the proper LispEnv. Anyway, modified the code to detect this case and retrieve the previous module environment from the global module hash table.