Fix for Issue #126
Do not parse a dotted list form if the car portion of the list is empty. It would be nice if the error were a little more friendly, but this is an edge case anyway and at least the parser points to the exact portion of the code that caused the problem.