Issue #151 - Macro hygiene for ER and syntax-rules
Preserve macro hygiene when using code that contains explicit renaming macros contained within syntax-rules macros. The issue was that the syntax-rules system was not passing renamed variables across to the ER system, so an identifier could be renamed by syntax-rules but the ER macro would then have no knowledge of the rename and would be unable to use `rename` to make the identifier hygienic.